[CHANGE] ChapterNumber from int to float64

+++ models/chapter.go
@@ -7,10 +7,10 @@ import (
 type Chapter struct {
 	gorm.Model
 
-	Name        string `json:"name"`
-	Site        string `json:"site"`
-	Number      int    `json:"number"`
-	PagesNumber int    `json:"pagesNumber"`
-	Pages       []Page `json:"pages"`
-	MangaID     int    `json:"mangaID"`
+	Name        string  `json:"name"`
+	Site        string  `json:"site"`
+	Number      float64 `json:"number"`
+	PagesNumber int     `json:"pagesNumber"`
+	Pages       []Page  `json:"pages"`
+	MangaID     int     `json:"mangaID"`
 }
